refactor(#135): rename files for lookup service provider list endpoint
